fix: refactor auth service FE
