+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/api/v2/controllers/OAuth2CallbackController.kt
@@ -37,7 +37,7 @@ class OAuth2CallbackController(
   ): String =
     "${tenant.authorizationUri}?" +
       "client_id=${tenant.clientId}&" +
-      "redirect_uri=${tenant.redirectUriBase + "/openId/auth_callback/" + tenant.domain}&" +
+      "redirect_uri=${tenant.redirectUriBase + "/openId/auth-callback/" + tenant.domain}&" +
       "response_type=code&" +
       "scope=openid profile email roles&" +
       "state=$state"
