+++ webapp/src/component/security/Sso/SsoRedirectionHandler.tsx
@@ -1,13 +1,11 @@
-import { FunctionComponent, useEffect } from 'react';
-import { Redirect, useHistory, useRouteMatch } from 'react-router-dom';
+import {FunctionComponent, useEffect} from 'react';
+import {Redirect, useHistory, useRouteMatch} from 'react-router-dom';
 
-import { LINKS, PARAMS } from 'tg.constants/links';
+import {LINKS, PARAMS} from 'tg.constants/links';
 
-import {
-  useGlobalActions,
-  useGlobalContext,
-} from 'tg.globalContext/GlobalContext';
-import { FullPageLoading } from 'tg.component/common/FullPageLoading';
+import {useGlobalContext,} from 'tg.globalContext/GlobalContext';
+import {FullPageLoading} from 'tg.component/common/FullPageLoading';
+import {useSsoService} from 'tg.component/security/SsoService';
 
 interface SsoRedirectionHandlerProps {}
 const LOCAL_STORAGE_STATE_KEY = 'oauth2State';
@@ -18,19 +16,17 @@ export const SsoRedirectionHandler: FunctionComponent<
   const allowPrivate = useGlobalContext((c) => c.auth.allowPrivate);
   const loginLoadable = useGlobalContext((c) => c.auth.authorizeOAuthLoadable);
 
-  const { loginWithOAuthCodeOpenId } = useGlobalActions();
+  const { loginWithOAuthCodeOpenId } = useSsoService();
   const match = useRouteMatch();
   const history = useHistory();
 
   useEffect(() => {
     const url = new URLSearchParams(window.location.search);
     const code = url.get('code');
-
     const state = url.get('state');
     const storedState = localStorage.getItem(LOCAL_STORAGE_STATE_KEY);
     if (storedState !== state) {
       history.replace(LINKS.LOGIN.build());
-      return;
     } else {
       localStorage.removeItem(LOCAL_STORAGE_STATE_KEY);
     }
